+++ core/src/sdk/interface/ConnectionImpl.cpp
@@ -89,10 +89,10 @@ ConnectionImpl::Insert(const std::string& table_name, const std::string& partiti
 }
 
 Status
-ConnectionImpl::Search(const std::string& table_name, const std::vector<std::string>& partiton_tags,
+ConnectionImpl::Search(const std::string& table_name, const std::vector<std::string>& partition_tags,
                        const std::vector<RowRecord>& query_record_array, const std::vector<Range>& query_range_array,
                        int64_t topk, int64_t nprobe, TopKQueryResult& topk_query_result) {
-    return client_proxy_->Search(table_name, partiton_tags, query_record_array, query_range_array, topk, nprobe,
+    return client_proxy_->Search(table_name, partition_tags, query_record_array, query_range_array, topk, nprobe,
                                  topk_query_result);
 }