+++ docker/deploy/gpu/ubuntu18.04/docker-entrypoint.sh
@@ -3,7 +3,7 @@
 set -e
 
 if [ "$1" == 'start' ]; then
-    cd /opt/milvus/scripts && ./start_server.sh
+    cd /var/lib/milvus/scripts && ./start_server.sh
 fi
 
 exec "$@"